Shopping page Skeleton
Still needs an array list with db connection to display the items from database. Just used some static data to get it running but the javascript works for sending data from shopping.html to the cart page. Could use as template for the product pages. Buttons on cart page do nothing at the moment. script is currently only in cart and shopping.html.